requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
python-dotenv==1.1.0
